body margin changed to 0 :
The reason they do this is so that if someone creates a website without CSS or if the CSS does not load, the website should still be readable. If the person who coded the HTML did so in good semantic order, the website should still be usable. If there was no default margin, then text would be hard to read because it would butt up against the edge of the browser. If the browser didn't have a default heading size, it would just look like normal text.